%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from set.default.arguments.R
%%
\name{control.gev2.default}
\alias{control.gev2.default}
\alias{inla.control.gev2.default}
\alias{inla.set.control.gev2.default.default}
\alias{set.control.gev2.default.default}
\alias{control.gev2.default.default}
\title{Control variables in control.gev2.default}
\description{Control variables in \code{control.gev2.default} for use in \code{inla}}
\usage{
inla.set.control.gev2.default.default(...)
control.gev2.default(beta.ab, q.location, q.mix, q.spread)
}
\arguments{
\item{...}{Possible arguments}
\item{q.location}{  The quantile level for the location parameter}
\item{q.spread}{  The quantile level for the spread parameter (must be < 0.5)}
\item{q.mix}{  The lower and upper quantile level for the mixing function}
\item{beta.ab}{  The parameters a and b in the Beta mixing function}
}
\value{
  The \code{control.gev2}-list is set within the corresponding \code{control.family}-list as control parameters to the \code{family="gev2"}
The function \code{control.gev2.default} is used to TAB-complete arguments and returns a list of given arguments.
The function \code{inla.set.control.gev2.default.default} returns a list with all the default values of all parameters within this control statement.
}
\seealso{
\code{\link{control.update}}, \code{\link{control.lincomb}}, \code{\link{control.group}}, \code{\link{control.mix}}, \code{\link{control.link}}, \code{\link{control.expert}}, \code{\link{control.compute}}, \code{\link{control.gev2.default}}, \code{\link{control.family}}, \code{\link{control.fixed}}, \code{\link{control.inla}}, \code{\link{control.predictor}}, \code{\link{control.results}}, \code{\link{control.mode}}, \code{\link{control.hazard}}, 
\code{\link{inla}}
}
