%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from sensitivity.R
%%
\name{inla.sens}

\title{Calculate sensitivity measurements}

\description{TODO}
\usage{
   inla.sens(inlaRes)
}

\arguments{
  \item{inlaRes}{Object returned by \code{inla} function.}
}
\value{
  \code{inla.sens}  plots robustness and returns object with different robustnesses
}
\author{Geir-Arne Fuglstad \email{geirarne.fuglstad@gmail.com}}
\examples{
 TODO
}
